+++ embeddable-say-editor/app/components/toolbar.gts
@@ -6,7 +6,7 @@ import type { SayController } from '@lblod/ember-rdfa-editor';
 import { get } from '@ember/helper';
 import type { TOC } from '@ember/component/template-only';
 import type { ResolvedPNode } from '@lblod/ember-rdfa-editor/utils/_private/types';
-import type { EditorSetup } from '../setup/setup-plugins';
+import type { EditorSetup } from '../plugins/setup/setup-plugins.ts';
 import type {
   ToolbarConfig,
   ToolbarGroupConfig,
