+++ embeddable-say-editor/app/plugins/setup/setup-plugins.ts
@@ -180,3 +180,5 @@ export function setupPlugins(args: PluginInitArgs): EditorSetup {
   }
   return result;
 }
+
+export type InitializedPluginName = keyof PluginSpecs;
