+++ embeddable-say-editor/package.json
@@ -67,10 +67,11 @@
     "@eslint/js": "^9.23.0",
     "@glimmer/component": "^2.0.0",
     "@glimmer/tracking": "^1.1.2",
-    "@glint/core": "1.4.1-unstable.c6fbad4",
-    "@glint/environment-ember-loose": "1.4.1-unstable.c6fbad4",
-    "@glint/environment-ember-template-imports": "1.4.1-unstable.c6fbad4",
-    "@glint/template": "1.4.1-unstable.c6fbad4",
+    "@glint/core": "1.4.1-unstable.afcad36",
+    "@glint/tsserver-plugin": "1.4.1-unstable.afcad36",
+    "@glint/environment-ember-loose": "1.4.1-unstable.afcad36",
+    "@glint/environment-ember-template-imports": "1.4.1-unstable.afcad36",
+    "@glint/template": "1.4.1-unstable.afcad36",
     "@lblod/ember-environment-banner": "^0.6.0",
     "@lblod/ember-rdfa-editor": "^12.1.0",
     "@lblod/ember-rdfa-editor-lblod-plugins": "^29.1.1",
