+++ test-angular/src/app/sayEditor.component.ts
@@ -1,5 +1,6 @@
 import { afterRender, Component, ElementRef } from '@angular/core';
 import { EditorElement, renderEditor } from '@lblod/embeddable-say-editor';
+import { EditorSetup } from '@lblod/embeddable-say-editor/app/plugins/setup/setup-plugins';
 import '@lblod/embeddable-say-editor/style.css';
 
 @Component({
@@ -21,7 +22,6 @@ export class SayEditor {
           title: 'angular embedded say editor',
           width: '1000px',
           height: '500px',
-          plugins: ['citation'],
         });
         this.editor?.setHtmlContent('Hello Angular');
       },
